+++ qpbenchmark/run.py
@@ -68,11 +68,7 @@ def run(
         for solver in filtered_solvers:
             for settings in filtered_settings:
                 time_limit = test_set.tolerances[settings].runtime
-                if (problem.name, solver) in test_set.known_solver_issues:
-                    logging.info(
-                        f"Skipping {problem.name} with {solver} "
-                        "as a known solver issue..."
-                    )
+                if test_set.skip_solver_issue(problem, solver):
                     failure = (
                         problem,
                         solver,